Move link-in-bio setup form to own component
